+++ WooCommerce/woocommerce_rms_normal/custom-gateway.php
@@ -1,0 +1,102 @@
+<?php
+/**
+ * Fiuu WooCommerce Shopping Cart Plugin
+ * 
+ * @author Fiuu Technical Team <technical-sa@razer.com>
+ * @version 3.1.0
+ * @example For callback : http://shoppingcarturl/?wc-api=WC_Molpay_Gateway
+ * @example For notification : http://shoppingcarturl/?wc-api=WC_Molpay_Gateway
+ */
+
+/**
+ * Plugin Name: WooCommerce Fiuu Normal
+ * Plugin URI: https://github.com/RazerMS/WordPress_WooCommerce_WP-eCommerce_ClassiPress
+ * Description: WooCommerce Fiuu | The leading payment gateway in South East Asia Grow your business with Fiuu payment solutions & free features: Physical Payment at 7-Eleven, Seamless Checkout, Tokenization, Loyalty Program and more for WooCommerce
+ * Author: Fiuu Tech Team
+ * Author URI: https://merchant.razer.com/
+ * Version: 3.1.0
+ * License: MIT
+ * Text Domain: wcmolpay
+ * Domain Path: /languages/
+ * For callback : http://shoppingcarturl/?wc-api=WC_Molpay_Gateway
+ * For notification : http://shoppingcarturl/?wc-api=WC_Molpay_Gateway
+ * Invalid Transaction maybe is because vkey not found / skey wrong generated
+ */
+
+/**
+ * If WooCommerce plugin is not available
+ * 
+ */
+function wcmolpay_woocommerce_fallback_notice() {
+    $message = '<div class="error">';
+    $message .= '<p>' . __( 'WooCommerce Fiuu Gateway depends on the last version of <a href="http://wordpress.org/extend/plugins/woocommerce/">WooCommerce</a> to work!' , 'wcmolpay' ) . '</p>';
+    $message .= '</div>';
+    echo $message;
+}
+
+//Load the function
+add_action('plugins_loaded', 'woocommerce_myplugin', 0);
+function woocommerce_myplugin(){
+    if ( !class_exists( 'WC_Payment_Gateway' ) ) {
+        add_action( 'admin_notices', 'wcmolpay_woocommerce_fallback_notice' );
+        return;
+    }
+
+    //Load language
+    load_plugin_textdomain( 'wcmolpay', false, dirname( plugin_basename( __FILE__ ) ) . '/languages/' );
+
+    add_filter( 'woocommerce_payment_gateways', 'wcmolpay_add_gateway' );
+
+    /**
+     * Add Fiuu gateway to ensure WooCommerce can load it
+     * 
+     * @param array $methods
+     * @return array
+     */
+    function wcmolpay_add_gateway( $methods ) {
+        $methods[] = 'WC_Molpay_Gateway';
+        return $methods;
+    }
+
+    include(plugin_dir_path(__FILE__) . 'class-gateway.php');
+}
+
+/**
+ * Custom function to declare compatibility with cart_checkout_blocks feature 
+*/
+function declare_cart_checkout_blocks_compatibility() {
+    // Check if the required class exists
+    if (class_exists('\Automattic\WooCommerce\Utilities\FeaturesUtil')) {
+        // Declare compatibility for 'cart_checkout_blocks'
+        \Automattic\WooCommerce\Utilities\FeaturesUtil::declare_compatibility('cart_checkout_blocks', __FILE__, true);
+    }
+}
+// Hook the custom function to the 'before_woocommerce_init' action
+add_action('before_woocommerce_init', 'declare_cart_checkout_blocks_compatibility');
+
+// Hook the custom function to the 'woocommerce_blocks_loaded' action
+add_action( 'woocommerce_blocks_loaded', 'oawoo_register_order_approval_payment_method_type' );
+
+/**
+ * Custom function to register a payment method type
+
+ */
+function oawoo_register_order_approval_payment_method_type() {
+    // Check if the required class exists
+    if ( ! class_exists( 'Automattic\WooCommerce\Blocks\Payments\Integrations\AbstractPaymentMethodType' ) ) {
+        return;
+    }
+
+    // Include the custom Blocks Checkout class
+    require_once plugin_dir_path(__FILE__) . 'class-block.php';
+
+    // Hook the registration function to the 'woocommerce_blocks_payment_method_type_registration' action
+    add_action(
+        'woocommerce_blocks_payment_method_type_registration',
+        function( Automattic\WooCommerce\Blocks\Payments\PaymentMethodRegistry $payment_method_registry ) {
+            // Register an instance of WC_Molpay_Gateway_Blocks
+            $payment_method_registry->register( new WC_Molpay_Gateway_Blocks );
+        }
+    );
+}
+?>
